requirements.txt
psycopg2
sqlalchemy
python-constraint
mlxtend
matplotlib
numpy
pytest
